requirements.txt
flask
gunicorn
google-genai
python-dotenv
pysqlite3
requests